Fix directory contents rendering
Previously, the elements in the directory contents list were constrained to a static height. This resulted in them overlapping each other if the window presenting them was constrained horizontally to a width smaller than the longer elements of the list. I was able to fix that issue by calculating an appropriate width in the CSS. The changes in this commit specifically target a follow-up problem: the directory and document icons were vertically centered with respect to the entire link text, thus making it difficult to identify the start of the link text.

By adding `.directory-contents-list-item-icon` and `.directory-contents-list-item-link`, the icons are able to be centered with respect to the first line of the link text. This is preferable to being centered with respect to the full height of the link since it allows the reader to direct their eyes to the start of the line rather than the vertical middle of the link text. I stole the approach from [this Stack Overflow answer](https://stackoverflow.com/a/56482928/6073927). It works in my case since I am setting a static height for both the directory and document icons.